pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "groq>=1.1.1",
    "llama-index>=0.14.18",
    "llama-index-embeddings-huggingface>=0.7.0",
    "llama-index-embeddings-openai>=0.6.0",
    "llama-index-graph-stores-neo4j>=0.7.0",
    "llama-index-llms-openai>=0.7.2",
    "llama-index-llms-openai-like>=0.7.1",
    "llama-index-readers-file>=0.6.0",
    "llama-index-readers-papers>=0.5.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
